Removes execute permissions from client and test sources.
Since in-tree testing is intended to use the build-tree copies of the
programs, removing execute permissions from the source-tree copies
avoids accidentally running the wrong copies.  This doesn't affect the
build-tree copies or the installed copies, since the permissions are
explicitly set as those are generated.  Not running the source copies
also avoids polluting the source directory with .pyc/.pyo files.

TESTED:
Verified correct permissions in all places.