systemd: Remove extra dependencies
The ntpviz services should not Requires=ntpd.service.  That has the
effect of starting ntpd.service (if it was stopped) whenever the timer
fires.  The same applies to ntploggps and gpsd.service.  Instead, we
should use Requisite.  We also need to do the same for ntp-wait.service;
while it has no timer, having a Requires=ntpd.service there ends up
causing the same problem for the ntpviz-* services by way of indirect
dependencies; this may be a systemd bug.

Also, instead of having timers.target pull in the timers, we have the
relevant daemon server (ntpd.service or gpsd.service) pull in the
timers.  This way, if the daemon is not running, the timers don't run
either.  This is especially important for gpsd.service.  If one doesn't
have gpsd installed, ntploggps.timer / ntploggps.service would fail.