gpsd: Fix non-updating of some derived values.
The gps_merge_fix() and gpsd_error_model() functions interact in a way
that prevents some derived values from being calculated more than
once.  This is because gpsd_error_model() only updates values which
haven't been set (presumably by the driver), but gps_merge_fix() never
copies unset values from the new fix data.  The net result is that
values not provided by the driver are only set the first time, and
left alone thereafter.

The fix here is to base the decisions on the new fix being merged,
where the driver's lack of provision of the values can be observed
correctly.

This problem was originally noticed in connection with the epx, epy,
and epc values, so the logic update is applied here to those values.
Since eps is computed from epx and epy, it is also similarly updated.
Other values may also need this treatment, but that is left for later.

Aside from impacting the reported epx, epy, epc, and epc values as
expected, this also affects synthetic GPGBS sentences in some ublox
cases, where the epx and epy values are included.  In addition, one
single case of an epd value being newly reported is observed,
presumably due to the changed epx/epy values impacting the filtering
decision.

TESTED:
The values displayed by xgps are now more plausible and not "stuck".
Except as noted above, regression differences are limited to the
expected four fields, with the new values being more plausible.