gps.h: Move gps_data_t.status to gps_fix_t.status
This allows the status (kinda sorta FAA status) to be merged like the
fix mode in gps_merge_fix().  This cleans up quirks in reporting of
RTK, DGPS, etc.  Most improvement in Trimble TSIP and NMEA drivers, but
changes in all drivers.

This changes the binary data structure, so not binary compatible.