SConstruct: Remove old packet.so if present.
Python (at least under some circumstances) prefers the old-style
packet.so to the new FFI-based version, leading to confusion.
The old packet.so needs to be removed to avoid this, both from
the build tree and from the install location.

TESTED:
Ran both build and install with old packet.so files present.
Verified that they were removed.