driver_oncore.c: only set LATLON, SPEED, TRACK mask and values if 2D fix, and ALT mask and value of 3D fix.
This changes the capture log check file for
oncore-binary-no-utc-offset.log, which has no fix and thus should not
give these values to gpsd.

The capture log oncore-binary-pos-hold.log also looses the position data
for the first report cycle, since the first At (position hold mode)
message has not arrived yet, and the driver therefore cannot verify that
the device really is in position-hold mode, and not just acquiring data.