driver_oncore.c: do not update the gpsd satellite information from the @@Bb message; only do the update in the @@Ea message.
The code for updating the satellite info was already complete in the
@@Ea message handling, including the use of the info stashed from @@Bb
in session->driver.oncore. Therefore this is only removing code.

Note that @@Bb messages are only sent from the device every few seconds.

Capture log .chk changes: the updates that came due to the @@Bb messages
are removed, the remaining updates are unchanged.

(Looking in the git history, the code for both the @@Bb and @@Ea
satellite info was added in commit 8a1904a055, and has except for
whitespace changes and changes of variable names not been changed
since.)