driver_ubx.c: Enable PROTVER detection for additional devices
Current code looks for PROTVER only in UBX-MON-VER.swVersion and
UBX-MON-VER.hwVersion fields. However, many devices carry the PROTVER
information in one of the UBX-MON-VER.extension fields.

E.g. ubxtool decoding of UBX-MON-VER from a recent mail on the gpsd-dev list*:

[...]
UBX-MON-VER:
  swVersion 2.01 (75331)
  hwVersion 00080000
  extension PROTVER 15.00
  extension GPS;SBAS;GLO;BDS;QZSS
[...]

The present patch looks in the UBX-MON-VER.extension fields for PROTVER
information if none was found in swVersion or hwVersion.

* https://lists.nongnu.org/archive/html/gpsd-users/2020-05/msg00066.html

Signed-off-by: Gary E. Miller <gem@rellim.com>