SConstruct: Add dependencies for Python imports.
Since Python import dependencies aren't discovered automatically, this
adds them explicitly, so that building any specific target also build
the proper dependencies.  This is redundant but harmless for
non-genereated files, and including non-generated files avoids the
need for this portion of the script to know which sources are
generated and which aren't.

TESTED:
All individual python-versions targets now work when starting from a
clean repo.