driver_oncore: Fix lat/lon/alt accuracy.
This fixes a regression failure which only appeared in 32-bit OpenBSD,
but it turns out that that was the only case where the behavior was
correct.  It affected both the oncore-binary and oncore-pos-hold
cases, and wasn't discovered until those test cases were added (and
the tests were run on 32-bit OpenBSD).

The basic problem is that the scale factors for converting the
lat/lon/alt parameters to doubles were specified as floats rather than
doubles, reducing the accuracy of the conversion.  Adjoining are three
other parameters, speed/track/dop, where the results are floats and
hence floats make sense for the scale factors, but lat/lon/alt are
doubles and should have the more precise scale factors.

Correcting the problem caused the result to match what was being
computed by 32-bit OpenBSD all along.  The most likely explanation is
that that particular toolchain was ignoring the 'f' suffix and using
doubles, anyway.

The history seems to be that the 'f' suffixes were initially added in
only the appropriate cases by commit 8a1904a055, but some later rework
in commit 7c9d538c62 copied the 'f' suffixes to the inappropriate
cases.  This was all over 10 years ago, and went unnoticed due to the
lack of tests.

TESTED:
Ran "build-all check" on macOS 10.5-10.13 (including 10.5 PPC), Ubuntu
14, CentOS 7, Fedora 25, FreeBSD 10.3, OpenBSD 5.6 (32- and 64-bit),
and NetBSD 6.1.5.