driver_nmea0183: Rework lat/lon decoding.
This reworks the decoding of the DDDMM.mmmmmmm lat/lon fields to be a
bit faster and a bit more accurate.  This was originally motivated by
failures on 32-bit OpenBSD with nmea-rtk.log and its derivative
nmea-fuzzy-cases.log.  This change alone doesn't completely fix that
problem, but it's still an improvement, and a followup change fixes
that as well.

Aside from OpenBSD-32, this change causes slight differences in a few
cases in the aforementioned two logfiles, all of which are single-unit
changes in low-order digits.  The two .chk files are updated
accordingly.

TESTED:
Ran "build-all check" on macOS 10.5-10.13 (including 10.5 PPC), Ubuntu
14, CentOS 7, Fedora 25, FreeBSD 10.3, OpenBSD 5.6 (32- and 64-bit),
and NetBSD 6.1.5, with the two .chk files updated as stated above.
All tests pass except for some OpenBSD-32 failures on those same two
logfiles, which are not yet fixed.