driver_nmea0183: Fix recent regressions on 32-bit OpenBSD
The issue with OpenBSD 5.6 is that the 32-bit version is set up to use
FLT_EVAL_METHOD = 2, which is a bit more accurate in principle but
also inconsistent with other platforms, including 64-bit OpenBSD 5.6.
The fix is to use a volatile variable for an intermediate value,
forcing it to be limited to a double.

Relying on exact comparisons of floating-point results is always
somewhat iffy, and the only reason it doesn't fail completely is due
to the consistent use of IEE-754 floating-point formats on all
platforms of interest.  It would be desirable to use fuzzy comparisons
on floating-point results, but that would require a major overhaul of
the test framework.

TESTED:
Ran "build-all check" on macOS 10.5-10.13 (including 10.5 PPC), Ubuntu
14, CentOS 7, Fedora 25, FreeBSD 10.3, OpenBSD 5.6 (32- and 64-bit),
and NetBSD 6.1.5.  All tests now pass on all platforms.