fake.py: Reduce default WRITE_PAD values.
Some other work may have reduced the need for WRITE_PAD.  Although it
doesn't appear that WRITE_PAD can currently be eliminated in non-Linux
cases, reducing it is worth a try.  This change:

1) Adds an explicit entry for OpenBSD, rather than having it rely on
the catch-all entry as it did previously.

2) Leaves the NetBSD 5 entry alone at its absurdly high 200ms value,
due to the lack of a NetBSD 5 system to test with.

3) Leaves the catch-all entry at 4ms.

4) Reduces all other nonzero entries to 1ms.

TESTED:
Ran "build-all check" on macOS 10.5-10.13 (including 10.5 PPC), Ubuntu
14, CentOS 7, Fedora 25, FreeBSD 10.3, OpenBSD 5.6 (32- and 64-bit),
and NetBSD 6.1.5.