gpsdclient.c: Fix fallthrough warning with some clang versions.
Correctly suppressing the warning across all compiler versions is
tricky, so this just eliminates the fallthrough by duplicating two
lines of code.  This particular warning seems to be limited to C++.

TESTED:
Ran "build-all check" with a couple of different clang and gcc
versions, avoiding the known issue with the Qt build and some clang
versions.