xgps: Fix crash when GST or AIS data present.
The code was using an obsolete method to turn on the checkboxes in
repsonse to GST or AIS data.  This went unnoticed until it was
run with a version of gpsd offering such data, which caused a crash.
The code now remembers the relevant objects instead of trying
to (unsuccessfully) look them up later.

TESTED:
No longer crashes when run with gpsd reporting GST.