SConstruct: Make python_shebang more convenient.
This now allows specifying just the program name, as an alternative to
the full path.  Values without a leading slash automatically have
"/usr/bin/env " prepended.  The default is now simply 'python', which
has the same result as before, but with a better example in the help
text.

It's not clear that making python_shebang and target_python
independent is either useful or convenient, but that issue isn't
addressed here.  At least it's now valid to set them to the same
value in all cases.

TESTED:
Default, program, and path cases behave as expected.