SConstruct: Add workaround for SCons 2.3.0 symlink bug.
The versioned library feature was just added in SCons 2.3.0, but it
had a bug where it was unable to overwrite existing symlinks.  This
was fixed in 2.3.2, but we allow 2.3.0.

The bug affected both rebuilds, which overwrite symlinks in the build
directory, and installs, where it conspires with another (still
current) bug that fails to remove library symlinks when uninstalling.

The fix is to monkey-patch os.symlink to remove the old symlink if
needed.  This is only applied when the SCons version is 2.3.0 or
2.3.1.

TESTED:
Symlink overwrites now work with SCons 2.3.0 on Ubuntu 14.
Also verified that os.symlink is not patched in SCons 3.1.2.