Revert "SConstruct: Make python_shebang more convenient."
This reverts commit e60579c437276e0e9857378d98d1cbecc2254517.

Many problems with it.

It imposes the constraint that the python_shebang must start with a
os.path.sep, a slash on Linux.  This prevents the use of many
valid shebangs.

Shebangs are defined in man execve(2).  They may start with white
space and the example at the top of the man page does so:

    #! interpreter [optional-arg]

The man page later says:

    "The interpreter must be a valid pathname for an executable which is
    not itself a script."

Nothing about it having to be an absolute path.

It also breaks shebangs with single quotes, double quotes, or Winblows
style:

    #!"C:\Python33\python.exe"

PEP 397 (https://www.python.org/dev/peps/pep-0397/)  is also on point.
Its first example uses white space, like this:

    #! interpreter [optional-arg]

PEP 397 also calls out as valid:

    #!python

So the patch that enforces the shabang to start with os.path.sep, or
have /usr/bin/env appended, does not conform to existing practice or
PEP 397,