address issue #102 fail when reopening ntrip caster ...
Woojin Jeong found when reconnecting to
ntrip://gnss:gnss@www.gnssdata.or.kr:2101/SUWN-RTCM31
that the reopening would fail due to the original pass butchering the URL argument.
JamesB192 suggested some diagnostics and then a warning triggering patch.
Woojin Jeong then countered with a line that fixed that patch.
Finally JamesB192 at the suggestion of gemiller packaged it as a merge request (in a simplified form)