gpsd.c: Fix check for invalid cycle times.
Commit e03fe942 changed the cycle times from double to timespec_t (which
in turn is a struct timespec). The values in struct timespec can be
negative.

The client can set the cycle time (which gets parsed from JSON into
devconf.cycle), the currently configured value is in
device->gpsdata.dev.cycle.

Before this commit, any positive value in devconf.cycle is leads to the
value being set to the currently configured value and no change happens.

After this commit, when devconf.cycle is zero or less it is set to the
currently configured values to prevent a change later on in the code,
otherwise devconf.cycle is later used to change the cycle time.

Signed-off-by: Gary E. Miller <gem@rellim.com>