SConscript: Work to always specify all dependencies.
scons does not like targets that come and go (if cleaning, if python,
etc). All targets are needed for proper cleaning. If a target should
not be built (if not python), then do not include the target in the
next layer sources.

scons gets confused by targets that are not a real file (shmclean, etc.).
Set them Pseudo (like in a Makefile) and use an Alias() for them.