systemd/gpsd.socket.in: Fix IPv6 wildcard binding.
The (commented) example line for a IPv6 wildcard socket was
incorrect, as it was bound to [::1] instead of [::].

Additionally, since systemd defaults to binding sockets to
both IPv6 and IPv4 unless the global system configuration has
been changed, including both wildcard sockets would result in
a startup failure stating 'address already in use'. The simple
fix is to explicitly bind the IPv6 wildcard socket to *only* IPv6,
so that the global system configuration is irrelevant.