SConscript: Fix for SCons 2.3.0.
SCons 2.3.0 lacks the Pseudo() method.  This change adds a simple
hack to make it a dummy when it's missing.  This doesn't seem to
cause trouble when building from clean, at least.

TESTED:
Verified (with pdb) that the dummy version is applied when needed, and
not when unneeded.
Ran "build-all check" on OSX 10.5-10.13, Ubuntu 14.04 (the problem
case), CentOS 7, Fedora 25, Debian 7, Debian 10, OpenBSD 5.6, and
NetBSD 6.1.5.