SConscript: Fix Python dependencies.
This updates the Python dependencies to correctly match the code.
Also updates the comments to reflect that fact that the dependencies
are now needed even for non-generated files.

The FIXME (now removed) was incorrect, but the updated comments may
clarify that.

Also tweaks python_progs into alphabetical order.

TESTED:
Built each Python program and module, verifying that the correct
dependencies were generated or copied.