SConscript: Add build-tree symlinks for 'gps' package.
The old symlink setup to allow accessing the 'gps' package from Python
programs in subdirectories was a casualty of the variantdir rework.
This resurrects it in a variantdir-compatible fashion, removing the
need to screw with PYTHONPATH to run such programs.

TESTED:
Ran 'build' and 'check' on macOS 10.5-10.15, Ubuntu 14.04, CentOS 7,
Fedora 25, Debian 7, Debian 10, FreeBSD 10.3, OpenBSD 5.6, and NetBSD
6.1.5.