SConscript: Define target aliases more consistently.
Due to the variantdir change, all targets defined by Utility() appear
to be located within the variantdir.  This was fixed in some cases by
adding Alias() calls, but this was not done in all cases.  This change
makes Alias() (and Pseudo()) part of Utility() itself, making the
behavior more consistent and eliminating the need for most of the
separate Alias() (and Pseudo()) calls.

The 'uninstall' target had a similar issue, which is also fixed.

TESTED:
Spot-checked several 'Utility()' targets, as well as the 'uninstall'
target.  They no longer need to be prefixed with the variantdir.