gps/packet.py: Rework packet library loader.
This reworks the search order for libgpsdpacket, as follows:

1) The directory which contains the packet.py's parent 'gps' package.
This is the location when running from the build tree, and is not
expected to work for installed code.

2) The library install directory.  This is the location used for
installed code.

3) The result of find_library() (as before).  This should ordinarily
not be needed.

Looking in the current directory was never appropriate, and (contrary
to the comments) doesn't match 'import' behavior, anyway.

Looking in GPSD_HOME was always inconsistent with the intended meaning
of GPSD_HOME.  It was somewhat useful as a workaround, but that should
no longer be the case.

TESTED:
Ran 'build' and 'install' on macOS 10.5-10.15, Ubuntu 14.04, CentOS 7,
Fedora 25, Debian 7, Debian 10, FreeBSD 10.3, OpenBSD 5.6, and NetBSD
6.1.5, verifying that gpscat and gpsfake could find the library, both
in the build tree and in the installed location.