SConscript:  Fix Python FFI install when shared=no.
Commit fc665ee22 fixed the building of the FFI library to make it
unconditionally shared, but neglected to similarly fix its install.
This meant that the library symlinks were never created in the
installed version.  It didn't really matter in practice, since the
lookup in packet.py uses the fully versioned name and doesn't require
the symlinks, but it still makes sense to install the symlinks for
consistency with other shared libraries.

TESTED:
The FFI library symlinks are now installed, even when shared=no.