SConscript: Create fixed-name symlink to build tree.
While using the version as the build-tree root is convenient for
building archives, it's inconvenient for scripts that need to access
the build products, as well as for helper symlinks.  To fix this, we
add a symlink from the old 'buildtmp' name to the versioned name.

TESTED:
The 'buildtmp' symlink is created, not included in the archives, and
removed by 'clean'.