SConscript: testbuild: inherit scons name and some options.
For the 'testbuild' target, this uses the possibly nonstandard name
of scons, and also inherits the -Q, -s, and -j options from the
main scons.

Also moves the 'scons_executable_name' setup earlier, so that it
doesn't look like it's specific to certain tests.

TESTED:
Ran 'testbuild' with a couple of option combinations.