gpsprof: Fix hang with 'instrumented' format.
Commit 65f4ae20 wasn't limited to doctsrings as the description
stated, but also added a (logically unnecessary) initialization of
'requires_time' to the 'plotter' base class.  However, this was added
as an instance variable, rather than a class variable as intended, so
it overrode the definitions in the subclasses.  The main effect was to
cause the 'instrumented' format to hang when it should have reported
"timing is not enabled".  Changing it to a class "variable" in the
base class fixes this.

Since 'requires_time' is actually a constant rather than a variable,
it should really be written in upper case, but that's left alone for
now.

Another effect of this bug was to allow the 'time' format to work and
provide what appears to be plausible output, rather than getting the
"timing is not enabled" error.  Whether this is appropriate is a
matter for further study.

TESTED:
Now gets "timing is not enabled" when expected.