SConscript: Fix build of www/cycle.svg.
The build of cycle.svg was using Utility(), which wasn't intended for
file targets, and failed in this case due to the Pseudo() call
(but only when 'dia' was installed).

Also, on the Mac (at least), running 'dia' launches the X11 server if
it isn't already running even though it's not actually displaying
anything.  That's fixed by clearing DISPLAY in its environment.  This
was previously being done for "xgps[speed] -V", for the same reason,
so this adjusts and reuses that code.

TESTED:
Building with and without 'dia' installed now works.
Building with 'dia' no longer launches X11 on the Mac.