SConscript: Accept asciidoctor26 as well
If asciidoctor is not found, look for asciidoctor26.  (There is a
general problem with languages with multiple not-quite-compatible
versions, leading to many packages needing to be installed for
multiple versions because some things that depend on them need one
version and some neeed another.  One approach ends up with version
suffixes.)

This shouldn't have any effect on people who have an "asciidoctor" in
their path, and be helpful for those who do not but have
asciidoctor26.

Fixes build of gpsd in a pkgsrc environment.