SConscript: work around inscrutable scons bug
After much gnashing of teeth, we have figured out that scons will
under some circumstances execute the install step for placing gps.h
into $includedir during the build phase.  To trigger this, it's
necessary that CPPFLAGS contain -I for the include directory, and
example1.c be included in htmlpages.  The theory is that scons somehow
guesses that compiling example1.c, in the presence of the -I argument
would reference the installed gps.h and therefore it should be
installed.  This is wrong because example1.c is not in fact built --
it's merely a text file to be in html as an example, the install
attempt is after everything else, and it's wrong to write outside of
the source directory during the build phase.

Lacking anyone having any idea about how to figure out why scons is
doing this and fix it, we add a workaround by changing the suffix of
the example not to end in .c.