libgps_dbus: Fix for platforms without clock_gettime().
Modules that use clock_gettime() need to include os_compat.h in order
to get the fallback support when needed.

Additionally, the current fallback only supports CLOCK_REALTIME, which
has been the only clock_gettime() type used so far.  In general, the
step adjustments which may be present in CLOCK_REALTIME are harmless
in the context of a timeout, so this just changes the new timeout to
use CLOCK_REALTIME.

TESTED:
Now builds on OSX 10.9.  Five-second dbus timeout works.