Removal of /usr/bin/python from Ubuntu forces python3 as a hard dependency.
Python 2 has been officially dead soince the first second of 2020, so
it's not too soon for this.  It wasn't forced until Ubuntu yanked the
binary in their latest upgade; besides them being the 800-pound
gorilla in the distro space, this certainly presages other major
distros doing the same thing.

It's not worth the effort to try to stay python2-friendly at this point.