use fgets to read an entire line at once
fread is subject to the MTU, which means it might not always grab
10000 bytes at once. This means you can likely get a partial read
of MTU minus header size, truncating the JSON data, and requiring
multiple reads, which are easy to get wrong. Instead, since we know
gpsd will always return a newline for each message, let PHP handle
it instead.