SConscript: Fix linking options with debug=yes.
Putting the -g3 only in CCFLAGS doesn't get it into the linking
command line.  This puts it in both LDFLAGS and LINKFLAGS, copying
the way the coverage options are handled.  The change has no effect
unless debugging is enabled.

TESTED:
On OSX, verified that debug=yes now gets debug symbols to lldb.