SConsript: Update set of ruby versions for asciidoctor
We look for e.g. asciidoctor31, if asciidoctor is not found, because
some packaging systems have multiple versions, because ruby is not
stable enough to only have one version.   This commit simply updates
the set of ruby versions looked for from 2.7, 2.6, to 3.1, 3.0, 2.7,
and should have no effect on those who have unadorned asciidoctor.