clients/ubxtools: fix raw recording
This fixes the raw recording on ubxtools. The cmdline option was existing but the raw file was never written. This commit fixes this by passing the raw file fd to the read function to record the info

A simple test: "ubxtool -R raw.txt" to record some data. Previously, the file was empty. Now, we can read back the file "ubxtool -f raw.txt".

Issue comes from the change in 88369fe2 that discarded the full opts from getting passed to the read() function.