gpsd/ntrip: detect and refuse chunked transfer-encoding
It looks like NTRIP v2.0 is heavily based on HTTP/1.1, with some casters
also using chunked transfers, as defined by RFC 9112, chap. 7.1, Chunked
Transfer Coding [0], like so:

    HTTP/1.1 200 OK\r\n
    [...headers...]\r\n
    Transfer-Encoding: chunked\r\n
    \r\n
    64\r\n
    64-bytes-worth-of-binary-message\r\n
    27\r\n
    27-bytes-worth-of-binary-message\r\n

This looks like what is actually causing issue #224 [1]: the caster for
the centipede network [2] does send such chunked replies, as illustrated
in the capture [3].

So far, we were assuming that the NTRIP stream was raw RTCM3 data, which
should be fed as is and unmolested to the GNSS receiver. The chunking
causes desynchronisation in the RTCM3 decoder, which detets invalide CRC
and finds garbage at the end of frames, becuse of the chunking markers.

Adding proper chunking support wil not be trivial, but we can at least
detect the situation and refuse to use such casters for now.

[0] https://datatracker.ietf.org/doc/html/rfc9112#name-chunked-transfer-coding
[1] https://gitlab.com/gpsd/gpsd/-/issues/224
[2] https://centipede.fr/
[3] https://gitlab.com/gpsd/gpsd/-/issues/224#note_1465246567

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>