gpsd/ntrip: update ntrip fd on connection update
When an NTRIP device is added dynamically (with gpsdctl), or when the
connection to the caster was lost and re-opened, the RTCM3 stream may
get lost and never forwarded to the local GNSS devices.

The root cause if that the filedescriptor used for the "established"
state of the connection may be different from the one we got when we
initially added the device.

Indeed, when we use gpsdctl to dynamically add a device, gpsd gets a
filedescriptor for that dealing with the control socket, then proceeds
to add the device requested via the control socket, while still having
the control socket opened. The device gets a filedescriptor for itself,
then gpsd replies to the control socket that the addition was OK, and
closes the control socket.

In the meantime, the NTRIP device proceeds with requesting the source
table. Then it closes the socket to the caster, and opens a new socket
to send the actual mountpoint request. This new socket may very well get
a filedescriptor that differs from the one used to request the source
table. Specifically, the fd from the control socket may get recycled.

In the case that another device gets removed at runtime, and the
connection to the caster drops (e.g. cellular glitch, WiFi roaming,
DSL lease renegotiation...), the new connection may also get an fd
different from the initial one (probably the fd from the removed device
gets recycled).

The issue arises also when there is a client that sporadically connects
to the daemon via the client socket, right when we want to reconnect the
NTRIP connection: the client got our old fd recycled, and our new
connection gets a different fd.

However, this change in filedescriptor is not tracked by gpsd (or any
other tool, but gpsd is where it matters), so the set of filedescriptors
to watch, all_fds, is not updated accordingly, so the new filedescriptor
will never be listened for, nor read from.

We fix all this by introducing a callback that a driver can call, to
inform of an fd closing or opening, so that it gets removed/added to the
set of watched fds.

We use that callback from the NTRIP net driver, to inform gpsd that the
fd to watch for has changed.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>