libgps/netlib.c: Rework enabling non-block and make binding configurable.
Instead of accepting SOCK_NONBLOCK as flags in netlib_connectsock1()
specify if the non-blocking mode should be enabled after or before
connect().

Also add a boolean parameter to the function to select between connect()
and bind() instead of hardcoding it for TCP vs UDP, which will allow
connecting to UDP ports in gps2udp.