gps/gps.py.in: no busy-waiting when reading from gpsd socket
ubxtool keeps one CPU 100% busy while it waits for data to read from the
gpsd socket. Running it under strace showed that it calls select() with
zero timeout in a loop:

...
11:02:34.049629 pselect6(4, [3], [], [], {tv_sec=0, tv_nsec=0}, NULL) = 0 (Timeout)
11:02:34.049649 pselect6(4, [3], [], [], {tv_sec=0, tv_nsec=0}, NULL) = 0 (Timeout)
11:02:34.049670 pselect6(4, [3], [], [], {tv_sec=0, tv_nsec=0}, NULL) = 0 (Timeout)
...

The busy waiting can be eliminated by passing the actual timeout value
to select(). In the reading loop in gps.py, the remaining time can be
easily calculated and passed as the argument to the self.ser.waiting()
function (which is basically a select() wrapper).

Fixing this problem exposed a bug in how the received bytes are decoded.
decode_func may not consume all input at once. Consumable input may be
left in self.out until decode_func returns zero, indicating that it
could not process any more input. So decode_func must be called in a
loop each time a buffer is received from the socket. The busy waiting
was hiding this issue, because decode_func was being called all the
time.

The "elif self.input_is_device:" branch probably needs similar
treatment, but I am testing only the gpsd usecase.