SConscript: provide variable names to MergeFlags
If the scons MergeFlags() function is provided with a list of
compiler/linker options, it needs to guess what options belong where. If
it doesn't recognize an option, it is silently ignored. There are also
ambiguous options that could be both in CFLAGS and LINKFLAGS (e.g. -spec).

Provide MergeFlags() with a dict instead of list to avoid the guesswork
in order to pass all options, even if some are not recognized by scons.
Switch LDFLAGS to LINKFLAGS for better compatibility with existing
scripts.