gpsmon/monitor_nmea0183.c: Fix gpsmon crash in NMEA.
When using FORTIFY_SOURCE the cc can create invalid code,  Avoid
the condition.  Also remove a lot of strnlen(field[0]) and a
slight refactor.

The code that compiles wrong:

    char buf[10];
    strlcpy(&buf[2], "src", sizeof(buf))

The size argument is larger than the remaining space in buf, which
should be OK since "src" still fits, but it creates a runtime error
anyway.