Sconscript+libgps: Restore Qt4 option.
The changes to allow Qt5 and Qt6 broke the previously working Qt4
case.  This is a minimal fix to restore the Qt4 option, at least when
specified explicitly.

The change in gpsutils for Qt6 was invalid for Qt<5.8.  It is now
conditional on the Qt version.

TESTED:
Now handles Qt4 properly, while still autodetecting Qt5/Qt6 properly.
Builds successfully with Qt4, Qt5.7, and Qt5.8.