gpsd: Just warn on non-root droproot error.
When (sometimes successfully) running as non-root, dropping root is
neither necessary nor legal, and the attempt to do so gets an EPERM
error.  This reports that case as just a warning, rather than a bogus
error.

Also fixes the indentation in the call.

TESTED:
No longer reports this as an error in non-root case.
Still reports it as a warning in non-root case.