gpsd: Just warn on unnecessary chmod() failure.
When a terminal has been successfully opened as non-root, adjusting
its permissions to allow later reopening as non-root is neither
necessary nor legal, and the attempt to do so gets an EPERM error.
This reports that case as just a warning, rather than a bogus error.

Also fixes the incorrect message text.

TESTED:
No longer reports this as an error in non-root case.
Still reports it as a warning in non-root case.